+++ debug_toolbar/middleware.py
@@ -80,7 +80,7 @@ class DebugToolbarMiddleware:
 
     def __call__(self, request):
         # Start tracking middleware execution time
-        middleware_start_time = perf_counter()
+        toolbar_start_time = perf_counter()
         # Decide whether the toolbar is active for this request.
         if self.async_mode:
             return self.__acall__(request)
@@ -101,16 +101,14 @@ class DebugToolbarMiddleware:
             # regardless of the response. Keep 'return' clauses below.
             for panel in reversed(toolbar.enabled_panels):
                 panel.disable_instrumentation()
-            middleware_end_time = perf_counter()
-            toolbar.middleware_time = (
-                middleware_end_time - middleware_start_time
-            ) * 1000
+            toolbar_end_time = perf_counter()
+            toolbar.toolbar_time = (toolbar_end_time - toolbar_start_time) * 1000
 
         return self._postprocess(request, response, toolbar)
 
     async def __acall__(self, request):
         # Start tracking middleware execution time
-        middleware_start_time = perf_counter()
+        toolbar_start_time = perf_counter()
         # Decide whether the toolbar is active for this request.
         show_toolbar = get_show_toolbar()
         if not show_toolbar(request) or DebugToolbar.is_toolbar_request(request):
@@ -134,10 +132,8 @@ class DebugToolbarMiddleware:
             # regardless of the response. Keep 'return' clauses below.
             for panel in reversed(toolbar.enabled_panels):
                 panel.disable_instrumentation()
-            middleware_end_time = perf_counter()
-            toolbar.middleware_time = (
-                middleware_end_time - middleware_start_time
-            ) * 1000
+            toolbar_end_time = perf_counter()
+            toolbar.middleware_time = (toolbar_end_time - toolbar_start_time) * 1000
 
         return self._postprocess(request, response, toolbar)