+++ docs/changes.rst
@@ -9,7 +9,7 @@ Pending
 * Converted to Django Commons PyPI release process.
 * Fixed a crash which occurred when using non-``str`` static file values.
 * Documented experimental async support.
-* Added middleware time to the timer panel.
+* Added toolbar time to the timer panel.
 
 5.0.0-alpha (2024-09-01)
 ------------------------